:seedling: Remove host logic from individual clusters
As it makes more sense to have a separate host inventory, we remove the
reference to the Hetzner cluster from the host object. This requires
some changes in the states, as we cannot continue with pre-processing,
before a bare metal machine consumes a host. Now all information and
credentials come directly or indirectly from the bare metal machine.
Therefore, the host waits in StateNone until a consumer ref is set.